doctest

+++ src/pyhf/probability.py
@@ -174,7 +174,7 @@ class Independent(_SimpleDistributionMixin):
             np.float64(-4.26248380...)
             >>> broadcast_value = pyhf.tensorlib.astensor([11.0])
             >>> independent.log_prob(broadcast_value)
-            np.float(-4.34774364...)
+            np.float64(-4.34774364...)
 
         Args:
             value (:obj:`tensor` or :obj:`float`): The value at which to evaluate the distribution
